Move initializers to the that source file; Updated copyright so I never have to change it again
